Move BeamJob out of __init__.py
